[Bug #1208] Fix problem with imputed hours in project view
The problem was related with the usage of the wrong function. It was
caused because of there're 2 functions with the same name, but doing
different things depending on the type of argument.

FEA: ItEr75S04BugFixing

+++ navalplanner-webapp/src/main/java/org/navalplanner/web/planner/TaskElementAdapter.java
@@ -510,7 +510,7 @@ public class TaskElementAdapter {
                 GanttDate result = null;
                 if (!(taskElement instanceof TaskGroup)) {
                     result = calculateLimitDate(assignedEffort
-                            .toHoursAsDecimalWithScale(2));
+                            .toHoursAsDecimalWithScale(2).intValue());
                 }
                 if (result == null) {
                     Integer hours = taskElement.getSumOfHoursAllocated();
